+++ CHANGELOG.md
@@ -1,15 +1,33 @@
+## 2.0.0
+
+- **(Breakingchanges)**
+  - Renamed class to `IpExterno` instead of `ip_externo`
+  - Removed some methods: `ReadIP`, `getIP`, `getResponse` e `getBody`
+  - Now the lib consumes new API to get your external ip
+- Added tests and .travis
+
+### Breakingchanges
+
+```dart
+import 'package:ip_externo/ip_externo.dart';
+
+void yourFunction() async {
+  final yourIp = await IpExterno.ipAddress;
+}
+``` 
+
 ## 1.0.3
 
-- Correção texto de saída em getIP();
+- Fixed output text from getIP();
 
 ## 1.0.2
 
-- Correção README.md
+- Fixed README.md
 
 ## 1.0.1
 
-- Correção texto de saída em getIP();
+- Fixed output text from getIP();
 
 ## 1.0.0
 
-- Versão inicial, desenvolvido por Mateus Penha.
+- First version, developer by Mateus Penha.